# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Astervia_wacraft-core_src_common_model.DescriptiveError:
    properties:
      content:
        description: Additional information about the error provided by the service
          that generated it. E. g. if an ORM returns an error, it will be propagated
          here as the raw error.
      message:
        description: Short text describing the error as best as we can for humans.
        type: string
    type: object
...
  github_com_Astervia_wacraft-core_src_phone-config_entity.PhoneConfig:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      display_phone:
        description: Display phone number (e.g., +1234567890)
        type: string
      id:
        type: string
      is_active:
        type: boolean
      meta_app_secret:
        type: string
      name:
        description: Friendly name for this configuration
        type: string
      updated_at:
        type: string
      waba_account_id:
        description: WhatsApp Business Account ID
        type: string
      waba_id:
        description: Phone Number ID from Meta (unique when active)
        type: string
      webhook_verify_token:
        type: string
      workspace:
        $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_entity.Workspace'
      workspace_id:
        type: string
    type: object
  github_com_Astervia_wacraft-core_src_phone-config_model.CreatePhoneConfig:
    properties:
      access_token:
        type: string
      display_phone:
        type: string
      is_active:
        type: boolean
      meta_app_secret:
        type: string
      name:
        type: string
      waba_account_id:
        description: WhatsApp Business Account ID
        type: string
      waba_id:
        description: Phone Number ID from Meta
        type: string
      webhook_verify_token:
        type: string
    required:
    - access_token
    - display_phone
    - meta_app_secret
    - name
    - waba_account_id
    - waba_id
    - webhook_verify_token
    type: object
  github_com_Astervia_wacraft-core_src_phone-config_model.UpdatePhoneConfig:
    properties:
      access_token:
        type: string
      display_phone:
        type: string
      is_active:
        type: boolean
      meta_app_secret:
        type: string
      name:
        type: string
      waba_account_id:
        type: string
      waba_id:
        type: string
      webhook_verify_token:
        type: string
    type: object
...
  github_com_Astervia_wacraft-core_src_user_entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        description: 'Deprecated: use workspace policies'
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Astervia_wacraft-core_src_workspace_entity.Workspace:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/github_com_Astervia_wacraft-core_src_user_entity.User'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: wacraft@astervia.tech
    name: Astervia Dev Team
    url: https://github.com/Astervia
  description: Backend server for the wacraft project. Handles WhatsApp Cloud API
    operations, including message sending, receiving, and webhook handling.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: wacraft Server API
  version: 0.1.0
paths:
...
  /workspace/{workspace_id}/phone-config:
    get:
      description: Returns a paginated list of phone configurations for the workspace.
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: created_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: created_at_geq
        type: string
      - in: query
        name: created_at_leq
        type: string
      - in: query
        name: display_phone
        type: string
      - description: The unique identifier
        in: query
        name: id
        type: string
      - in: query
        name: is_active
        type: boolean
      - default: 10
        description: Number of items to return
        in: query
        name: limit
        type: integer
      - in: query
        name: name
        type: string
      - default: 0
        description: The offset from where to start the items
        in: query
        name: offset
        type: integer
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: updated_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: updated_at_geq
        type: string
      - in: query
        name: updated_at_leq
        type: string
      - description: Phone Number ID
        in: query
        name: waba_id
        type: string
      - in: query
        name: workspace_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of phone configs
          schema:
            items:
              $ref: '#/definitions/github_com_Astervia_wacraft-core_src_phone-config_entity.PhoneConfig'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: List phone configurations
      tags:
      - Phone Config
    post:
      consumes:
      - application/json
      description: Creates a new phone configuration for the workspace.
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Phone config data
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/github_com_Astervia_wacraft-core_src_phone-config_model.CreatePhoneConfig'
      produces:
      - application/json
      responses:
        "201":
          description: Created phone config
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_phone-config_entity.PhoneConfig'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Create phone configuration
      tags:
      - Phone Config
  /workspace/{workspace_id}/phone-config/{id}:
    delete:
      description: Deletes a phone configuration from the workspace.
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Phone Config ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Delete phone configuration
      tags:
      - Phone Config
    get:
      description: Returns a specific phone configuration by ID.
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Phone Config ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Phone config
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_phone-config_entity.PhoneConfig'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Get phone configuration
      tags:
      - Phone Config
    patch:
      consumes:
      - application/json
      description: Updates an existing phone configuration.
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Phone Config ID
        in: path
        name: id
        required: true
        type: string
      - description: Phone config data
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/github_com_Astervia_wacraft-core_src_phone-config_model.UpdatePhoneConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Updated phone config
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_phone-config_entity.PhoneConfig'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Update phone configuration
      tags:
      - Phone Config
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"